Add files via upload
Stored Procedures
Stored Procedures are a way to save your SQL codes and can reuse it over and over again. When saving it, you can call that stored procedure and it will execute all the codes that you wrote within your stored procedure.
It is really helpful for storing complex queries, simplifying repetitive codes and just enhancing the overall performance.
When using a delimiter with dollar signs, you have to end with dollar signs too and therefore the semi-colon will no longer be a delimiter that tells us when is the end of a stored procedure.
Make sure to change the delimiter with a semi-colon afterwards otherwise you will keep sing the dollar signs. Delimiter;
Perimeters
Perimeters are variables that are passed as an input into a stored procedure. I
It allows the stored procedure to accept an input value and place it into your code.